+++ router.h
@@ -8,7 +8,7 @@ class Router
 public:
     Router(Graph graph, double cost, double passengerFee, double maxDistance);
     void buildRoute();
-    int chooseStop(std::vector<std::pair<int, int>> routes);
+    int chooseStop(std::vector<std::pair<int, double>> routes);
 private:
     Graph graph;
     double cost, passengerFee, maxDistance;
